# src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"AccountType": {
				"enum": [
					"CORRENTE",
					"POUPANCA"
				],
				"type": "string"
			},
			"AccountTransaction": {
				"properties": {
					"id": {
						"type": "string"
					},
					"bankAccountId": {
						"type": "string"
					},
					"transactionId": {
						"type": "string"
					},
					"fromAccount": {
						"$ref": "#/components/schemas/BankAccount"
					},
					"toAccount": {
						"$ref": "#/components/schemas/BankAccount"
					},
					"sentTransaction": {
						"$ref": "#/components/schemas/Transaction"
					},
					"receivedTransaction": {
						"$ref": "#/components/schemas/Transaction"
					}
				},
				"required": [
					"bankAccountId",
					"transactionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BankAccount": {
				"properties": {
					"id": {
						"type": "string"
					},
					"accountNumber": {
						"type": "string"
					},
					"accountType": {
						"$ref": "#/components/schemas/AccountType"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"sentTransactions": {
						"items": {
							"$ref": "#/components/schemas/AccountTransaction"
						},
						"type": "array",
						"default": []
					},
					"receivedTransactions": {
						"items": {
							"$ref": "#/components/schemas/AccountTransaction"
						},
						"type": "array",
						"default": []
					}
				},
				"required": [
					"id",
					"accountNumber",
					"accountType",
					"balance",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionType": {
				"enum": [
					"PIX",
					"DEPOSIT",
					"WITHDRAW"
				],
				"type": "string"
			},
			"Transaction": {
				"properties": {
					"id": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"$ref": "#/components/schemas/TransactionType"
					},
					"pixKey": {
						"type": "string"
					},
					"isReverted": {
						"type": "boolean"
					},
					"bankAccountFrom": {
						"$ref": "#/components/schemas/Partial_AccountTransaction_"
					},
					"bankAccountTo": {
						"$ref": "#/components/schemas/Partial_AccountTransaction_"
					}
				},
				"required": [
					"amount",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_AccountTransaction_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"bankAccountId": {
						"type": "string"
					},
					"transactionId": {
						"type": "string"
					},
					"fromAccount": {
						"$ref": "#/components/schemas/BankAccount"
					},
					"toAccount": {
						"$ref": "#/components/schemas/BankAccount"
					},
					"sentTransaction": {
						"$ref": "#/components/schemas/Transaction"
					},
					"receivedTransaction": {
						"$ref": "#/components/schemas/Transaction"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "bank-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/account-transactions": {
			"post": {
				"operationId": "CreateTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountTransaction"
								}
							}
						}
					}
				},
				"tags": [
					"Account Transactions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AccountTransaction"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AccountTransaction"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Account Transactions"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/bank-accounts": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BankAccount"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"accountType": {
										"$ref": "#/components/schemas/AccountType"
									},
									"accountNumber": {
										"type": "string"
									}
								},
								"required": [
									"accountType",
									"accountNumber"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BankAccount"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": []
			}
		},
		"/bank-accounts/{accountNumber}": {
			"get": {
				"operationId": "Read",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/BankAccount"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BankAccount"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"balance": {
										"type": "number",
										"format": "double"
									},
									"accountType": {
										"$ref": "#/components/schemas/AccountType"
									}
								},
								"required": [
									"balance",
									"accountType"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bank-accounts/by-id/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/BankAccount"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bank-accounts/{accountNumber}/deposit": {
			"post": {
				"operationId": "Deposit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BankAccount"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"amount"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bank-accounts/{accountNumber}/withdraw": {
			"post": {
				"operationId": "Withdraw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BankAccount"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"amount"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bank-accounts/search": {
			"post": {
				"operationId": "Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BankAccount"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"minBalance": {
										"type": "number",
										"format": "double"
									},
									"accountType": {
										"$ref": "#/components/schemas/AccountType"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bank-accounts/transaction": {
			"post": {
				"operationId": "CreateNewTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Transaction"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"type": {
										"$ref": "#/components/schemas/TransactionType"
									},
									"amount": {
										"type": "number",
										"format": "double"
									},
									"toAccountNumber": {
										"type": "string"
									},
									"fromAccountNumber": {
										"type": "string"
									}
								},
								"required": [
									"type",
									"amount",
									"toAccountNumber",
									"fromAccountNumber"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bank-accounts/{accountNumber}/recent-transactions": {
			"get": {
				"operationId": "GetRecentTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Transaction"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bank-accounts/transfer": {
			"post": {
				"operationId": "Transfer",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									},
									"toAccountNumber": {
										"type": "string"
									},
									"fromAccountNumber": {
										"type": "string"
									}
								},
								"required": [
									"amount",
									"toAccountNumber",
									"fromAccountNumber"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bank-accounts/{accountNumber}/transactions": {
			"get": {
				"operationId": "GetTransactionsForAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Transaction"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bank Accounts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/transactions": {
			"post": {
				"operationId": "CreateTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Transaction"
								}
							}
						}
					}
				},
				"tags": [
					"Transactions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"type": {
										"$ref": "#/components/schemas/TransactionType"
									},
									"amount": {
										"type": "number",
										"format": "double"
									},
									"toAccountNumber": {
										"type": "string"
									},
									"fromAccountNumber": {
										"type": "string"
									}
								},
								"required": [
									"type",
									"amount",
									"toAccountNumber",
									"fromAccountNumber"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Transaction"
								}
							}
						}
					}
				},
				"tags": [
					"Transactions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Transaction"
							}
						}
					}
				}
			}
		},
		"/transactions/{accountNumber}/recent": {
			"get": {
				"operationId": "GetRecentTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Transaction"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "accountNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/transactions/transfer": {
			"post": {
				"operationId": "TransferBetweenAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Transaction"
								}
							}
						}
					}
				},
				"tags": [
					"Transactions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									},
									"toAccountNumber": {
										"type": "string"
									},
									"fromAccountNumber": {
										"type": "string"
									}
								},
								"required": [
									"amount",
									"toAccountNumber",
									"fromAccountNumber"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}